+++ pyClarion/components/__init__.py
@@ -2,3 +2,4 @@
 
 from pyClarion.components.nacs import *
 from pyClarion.components.general import *
+from pyClarion.components.pull_rules import *